refactor: rename convertor to converter and merge modules
- Merge legacy GraphConvertor (CPU, pymatgen) and GPU BatchGraphConverter
  into a single converter.py module
- Rename GraphConvertor → GraphConverter (keep alias for backward compat)
- Delete convertor.py (all code now lives in converter.py)
- Rename test_convertor.py → test_converter.py
- Update build.py imports
